fix: workaround for file exists
Apparently this can't detect the existence of the symlink on other platform (e.g. Windows symlink on Linux) so we just forcefully delete it.

Not tested yet but it should work, if it doesn't then we'll have to fallback to "del" command on Windows and "rm" on Linux which works